synthetics monitor saved object
