+++ x-pack/plugins/uptime/server/lib/alerts/tls.test.ts
@@ -12,7 +12,7 @@ import { CertResult, DynamicSettings } from '../../../common/runtime_types';
 import { createRuleTypeMocks, bootstrapDependencies } from './test_utils';
 import { DYNAMIC_SETTINGS_DEFAULTS } from '../../../common/constants';
 
-import { savedObjectsAdapter, UMSavedObjectsAdapter } from '../saved_objects';
+import { savedObjectsAdapter, UMSavedObjectsAdapter } from '../saved_objects/saved_objects';
 
 /**
  * This function aims to provide an easy way to give mock props that will
