+++ x-pack/plugins/uptime/server/lib/alerts/tls_legacy.ts
@@ -18,7 +18,7 @@ import { ActionGroupIdsOf } from '../../../../alerting/common';
 import { AlertInstanceContext } from '../../../../alerting/common';
 import { AlertInstance } from '../../../../alerting/server';
 
-import { savedObjectsAdapter } from '../saved_objects';
+import { savedObjectsAdapter } from '../saved_objects/saved_objects';
 import { createUptimeESClient } from '../lib';
 import {
   DEFAULT_FROM,
