+++ x-pack/plugins/uptime/server/lib/lib.ts
@@ -16,7 +16,7 @@ import type * as estypes from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import { UMBackendFrameworkAdapter } from './adapters';
 import { UMLicenseCheck } from './domains';
 import { UptimeRequests } from './requests';
-import { savedObjectsAdapter } from './saved_objects';
+import { savedObjectsAdapter } from './saved_objects/saved_objects';
 import { ESSearchResponse } from '../../../../../src/core/types/elasticsearch';
 import { RequestStatus } from '../../../../../src/plugins/inspector';
 import { getInspectResponse } from '../../../observability/server';
